add comments, move _base files
