+++ src/viam/sdk/registry/registry.cpp
@@ -12,12 +12,12 @@
 
 #include <viam/api/robot/v1/robot.pb.h>
 
-#include <viam/sdk/components/component_base.hpp>
+#include <viam/sdk/components/component.hpp>
 #include <viam/sdk/components/generic/generic.hpp>
 #include <viam/sdk/resource/resource.hpp>
-#include <viam/sdk/resource/resource_base.hpp>
+#include <viam/sdk/resource/resource.hpp>
 #include <viam/sdk/resource/resource_type.hpp>
-#include <viam/sdk/services/service_base.hpp>
+#include <viam/sdk/services/service.hpp>
 
 namespace viam {
 namespace sdk {
