+++ src/viam/sdk/registry/registry.hpp
@@ -15,7 +15,7 @@
 
 #include <viam/sdk/config/resource.hpp>
 #include <viam/sdk/resource/resource.hpp>
-#include <viam/sdk/resource/resource_base.hpp>
+#include <viam/sdk/resource/resource.hpp>
 #include <viam/sdk/resource/resource_manager.hpp>
 #include <viam/sdk/resource/resource_server_base.hpp>
 #include <viam/sdk/resource/resource_type.hpp>
